requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
Flask==2.2.2
gevent==21.12.0
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
pycparser==2.21
python-dotenv==0.20.0
Werkzeug==2.2.2
zope.event==4.5.0
zope.interface==5.4.0
